+++ src/cpufreqd_acpi_battery.h
@@ -17,8 +17,8 @@
  *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
 
-int acpi_battery_init(void);
-int acpi_battery_exit(void);
+short int acpi_battery_init(void);
+short int acpi_battery_exit(void);
 int acpi_battery_parse(const char *ev, void **obj);
 int acpi_battery_evaluate(const void *s);
 int acpi_battery_update(void);
